Move Monty representation stuff to its own module.
